% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_bar.R
\name{plot_go_bar}
\alias{plot_go_bar}
\title{Plot GO Bar Chart}
\usage{
plot_go_bar(E, output_file, aspect_ratio = 2.5, height = 7)
}
\arguments{
\item{output_file}{A string specifying the path and filename for the output PDF file.}

\item{aspect_ratio}{A numeric value for the aspect ratio of the plot (default: 2.5).}

\item{height}{A numeric value for the height of the plot in inches (default: 7).}

\item{data}{A data frame prepared by the `prepare_go_data` function.}
}
\value{
Saves a PDF file with the generated bar chart.
}
\description{
Creates a bar chart to visualize GO enrichment results and saves it to a file.
}
