% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_go_data.R
\name{prepare_go_data}
\alias{prepare_go_data}
\title{Prepare GO Data}
\usage{
prepare_go_data(file_path)
}
\arguments{
\item{file_path}{A string specifying the path to the GO enrichment results file (TSV format).}
}
\value{
A data frame with the columns `genes_number`, `category`, and `description`, where
        `description` is ordered based on the `category`.
}
\description{
Reads and preprocesses a GO enrichment results file, filtering and formatting the data
for visualization.
}
